[mediasdk][codec] Fix Coverity INTEGER_OVERFLOW in BRC
Fix two signed integer overflow defects reported by Coverity static analysis:

CID 3515584 (umc_mpeg2_brc.cpp, CheckHRDParams):
  Cast mParams.HRDBufferSizeBytes to uint32_t before multiplying by 8
  to prevent signed 32-bit overflow.

CID 3544062 (umc_video_brc.cpp, CheckCorrectParams_MPEG2):
  Cast bufferSizeBytes to uint32_t before left-shifting by 3
  to prevent signed 32-bit left-shift overflow.

+++ _studio/shared/umc/codec/brc/src/umc_mpeg2_brc.cpp
@@ -240,9 +240,9 @@ Status MPEG2BRC::CheckHRDParams()
     }
   }
   mParams.HRDBufferSizeBytes = mHRD.bufSize / 8;
-  mHRD.bufSize = mParams.HRDBufferSizeBytes * 8;
+  mHRD.bufSize = (uint32_t)mParams.HRDBufferSizeBytes * 8;
   mParams.HRDInitialDelayBytes = (int32_t)(mHRD.bufFullness / 8);
-  mHRD.bufFullness = mParams.HRDInitialDelayBytes * 8;
+  mHRD.bufFullness = (double)mParams.HRDInitialDelayBytes * 8;
 
   if (mHRD.bufSize < mHRD.inputBitsPerFrame)
     return UMC_ERR_INVALID_PARAMS;
